+++ packages/examples/src/servers/json-server.ts
@@ -1,3 +1,6 @@
-import { runJsonServer } from 'monaco-languageclient-examples/node';
+import { resolve } from 'path';
+import { getLocalDirectory, runJsonServer } from 'monaco-languageclient-examples/node';
 
-runJsonServer();
+const baseDir = resolve(getLocalDirectory(import.meta.url));
+const relativeDir = '../../../../node_modules/monaco-languageclient-examples/dist/json/server/json-server.js';
+runJsonServer(baseDir, relativeDir);