+++ packages/monaco-editor-wrapper/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to npm module [monaco-editor-wrapper](https://www.npmjs.com/
 ## [2.3.0] - 2023-10-xy
 
 - Properly separate and define classic and extended editor [#54](https://github.com/TypeFox/monaco-components/pull/54)
-- Updated to `monaco-languageclient@6.5.2`
+- Updated to `monaco-languageclient@6.5.3`
 
 ## [3.2.5] - 2023-10-13